Improved early JS termination to require printability when printability constraints are present.
Declare 2D patterns always printable.

Fix warning in autocover_tools.py

+++ pattern_optimization/experiments/autocover/autocover_tools.py
@@ -67,8 +67,8 @@ def coverageExpansionOptimizer(config, lut, grid, constraints):
 
         # TODO: take N closest points that are within some minimal distance of each other.
         startLUTIndices = np.unique([order[0],
-                                     order[min(math.ceil(0.05 * len(order)), len(order) - 1)],
-                                     order[min(math.ceil(0.10 * len(order)), len(order) - 1)]])
+                                     order[min(int(math.ceil(0.05 * len(order))), len(order) - 1)],
+                                     order[min(int(math.ceil(0.10 * len(order))), len(order) - 1)]])
         if (config.get('singleClosestInit', False)):
             startLUTIndices = [order[0]]
         for i in startLUTIndices: opt.enqueueJob(targetE, targetNu, lut.params[i])
